Move `pkg/prometheus` to `pkg/metrics/prometheus`
