+++ pkg/dot/metrics/prometheus/settings.go
@@ -1,9 +1,9 @@
 package prometheus
 
 import (
+	prom "github.com/qdm12/dns/pkg/metrics/prometheus"
 	middleware "github.com/qdm12/dns/pkg/middlewares/metrics"
 	middlewarenoop "github.com/qdm12/dns/pkg/middlewares/metrics/noop"
-	prom "github.com/qdm12/dns/pkg/prometheus"
 )
 
 type Settings struct {
