+++ pkg/middlewares/metrics/prometheus/settings.go
@@ -1,7 +1,7 @@
 package prometheus
 
 import (
-	prom "github.com/qdm12/dns/pkg/prometheus"
+	prom "github.com/qdm12/dns/pkg/metrics/prometheus"
 )
 
 type Settings struct {
